==> Notebooks/0. Dataset Condensation.ipynb <==
import os
import pandas as pd
import re

debates_header = ["id", "text"]
meta_header = ["Text_ID","ID","Title","Date","Body","Term","Session","Meeting","Sitting","Agenda","Subcorpus","Lang","Speaker_role","Speaker_MP","Speaker_minister","Speaker_party","Speaker_party_name","Party_status","Party_orientation","Speaker_ID","Speaker_name","Speaker_gender","Speaker_birth"]

#Unione dei diversi file per ogni seduta in un unico file
def merge_files(regex_pattern, header):
    root_dir = "data/virgin_dataset"
    dfs = []
    n_sedute = 0
    
    for subdir, _, files in os.walk(root_dir):
        for file in files:
            if re.match(regex_pattern, file):
                n_sedute += 1
                file_path = os.path.join(subdir, file)

                if len(header) > 2: #if meta file
                    df = pd.read_csv(file_path, sep="\t", encoding="utf-8", names=header, header=0) 
                else:
                    df = pd.read_csv(file_path, sep="\t", encoding="utf-8", names=header)
                dfs.append(df)
    print("N_sedute: ", n_sedute)
    return dfs

debates_df = pd.concat(merge_files(r".*.txt", debates_header), ignore_index=True)

meta_df = pd.concat(merge_files(r".*meta.tsv", meta_header), ignore_index=True)

debates_df

meta_df

#### Creazione file

debates_df.to_csv("data/condensed_dataset/debates.tsv", index=False)

meta_df.to_csv("data/condensed_dataset/meta.tsv", index=False)

#### Creazione file merged

merged_df = meta_df.merge(debates_df, how='left', left_on='ID', right_on="id")

merged_df['text_length'] = merged_df['text'].str.len() #aggiungo una colonna con la lunghezza del testo

merged_df.to_csv("data/condensed_dataset/merged.tsv", index=False)

'''
Codice utile a risolvere un problema di annotazione del dataset: 
ogni record dei file meta terminava con un \t superfluo, che andava rimosso
perché altrimenti pandas vedeva una colonna in più

root_dir = "data/virgin_dataset"
dfs = []

for subdir, _, files in os.walk(root_dir):
    for file in files:
        if re.match(r".*meta.tsv", file):
            file_path = os.path.join(subdir, file)
            #print(file_path)

            # Preprocess the file to remove trailing tabs or spaces
            with open(file_path, "r", encoding="utf-8") as infile:
                lines = infile.readlines()

            # Strip trailing tabs/spaces from each line
            cleaned_lines = [line.replace('\t\n', '\n') for line in lines]

            # Save the cleaned file temporarily or overwrite it
            with open(file_path, "w", encoding="utf-8") as outfile:
                outfile.writelines(cleaned_lines)
'''
