==> Notebooks/1.1 Dataset Cleaning.ipynb <==
import pandas as pd

debates_df = pd.read_csv("data/condensed_dataset/debates.tsv", sep=",", encoding="utf-8")
meta_df = pd.read_csv("data/condensed_dataset/meta.tsv", sep=",", encoding="utf-8") 

for elem in set(meta_df["Party_orientation"]):
    print(elem, meta_df[meta_df["Party_orientation"]==elem].shape[0])

Quanti e quali valori distinti per ogni attributo dei metadati?

for att in list(meta_df):
    print("\n======", att, "======")
    dist_values = set(meta_df[att])
    if len(dist_values) < 150:
        print(dist_values)
    else:
        print("troppi valori distinti")
        

Elimino le colonne del dataset meta che non sono utili al nostro studio:
 - Text_ID: si può ricavare da ID
 - Title: non utile
 - Body: ha un solo valore distinto
 - Session: ha un solo valore distinto
 - Meeting: ha un solo valore distinto
 - Sitting: non utile
 - Agenda: ha un solo valore distinto
 - Lang: ha un solo valore distinto
 - Speaker_name: sovrapponibile a Speaker_ID

columns_to_delete = ["Text_ID", "Title", "Body", "Session", "Meeting", "Sitting", "Agenda", "Lang", "Speaker_name"]
selected_meta_df = meta_df.drop(columns=columns_to_delete)

selected_meta_df.head()

### Rimuovo i gruppi con meno di 2000 interventi

distinct_couples = set([tuple(x) for x in selected_meta_df[["Speaker_party", "Speaker_party_name"]].values])

distinct_couples

parties_to_drop = []
for party in distinct_couples:
    print("\n=====", party[0], "======")
    cnt = (selected_meta_df['Speaker_party'] == party[0]).sum()
    if cnt > 2000:
        print(cnt)
    else:
        print("troppi pochi interventi")
        parties_to_drop.append(party[0])
        
parties_to_drop.append("Misto")
parties_to_drop.append("-") #rimuovo anche i record non associati a un partito o al misto

parties_to_drop

final_meta_df = selected_meta_df.drop(selected_meta_df[selected_meta_df.Speaker_party.isin(parties_to_drop)].index)

selected_meta_df.shape

final_meta_df.shape

for att in list(final_meta_df):
    print("\n======", att, "======")
    dist_values = set(final_meta_df[att])
    if len(dist_values) < 150:
        print(dist_values)
    else:
        print("troppi valori distinti")

### Ultime rifiniture:
- elimino la colonna Speaker_MP perché adesso ha un solo valore unico
- elimino la tutte le righe dove Speaker_role = Presidente perché si tratta perlopiù di comunicazioni ufficiali. Elimino poi la colonna Speaker_role perché a quel punto avrà un solo valore unico

real_final_meta_df = final_meta_df.drop(final_meta_df[final_meta_df.Speaker_role == "Presidente"].index)
real_final_meta_df = real_final_meta_df.drop(columns=["Speaker_role", "Speaker_MP"])

real_final_meta_df.shape

### Adesso si fa il merge con debates_df, per proseguire la fase di cleaning. 
Si vorranno eliminare gli interventi troppo corti

merged_df = real_final_meta_df.merge(debates_df, how='left', left_on='ID', right_on="id")

merged_df = merged_df.drop(columns=["id"])

merged_df['text_length'] = merged_df['text'].str.len() #aggiungo una colonna con la lunghezza del testo

merged_df.text_length.describe()

merged_df[['text_length', "Speaker_party"]].groupby("Speaker_party").mean()

Elimino i testi meno lunghi di 4500 caratteri

merged_df[merged_df["text_length"]<4500]["ID"].count()

final_merged_df = merged_df.drop(merged_df[merged_df["text_length"]<4500].index)

final_merged_df.shape

final_merged_df[['text_length', "Speaker_party"]].groupby("Speaker_party").count()

#### Creazione file Cleaned 

final_merged_df.to_csv("data/cleaned_dataset/merged.tsv", index=False)

real_final_meta_df.to_csv("data/cleaned_dataset/meta.tsv", index=False)
