==> Notebooks/1.2 Dataset Advanced Cleaning.ipynb <==
import pandas as pd
import re

cleaned_df = pd.read_csv("data/cleaned_dataset/merged.tsv", sep=",", encoding="utf-8")

refined_df = cleaned_df
remove_list = ["(FdI)", "(M5S)", "(PD)", "(FI)", "(FI-BP)", "(PdL)", "(L-SP)", "(LN-Aut)", "(NCD)", "(SCpI)"]
refined_df["text"] = cleaned_df['text'].str.replace(r'\[\[.*?\]\]', '', regex=True) #rimuovo le stringe come [[Applausi]]
refined_df['text'] = refined_df['text'].replace('|'.join(map(re.escape, remove_list)), '', regex=True) #Rimuovo le note come (M5S)

diz = {
    "LN-Aut": "Lega",
    "L-SP": "Lega",
    "M5S.1": "M5S",
    "M5S.2": "M5S",
}
refined_df = refined_df.replace({"Speaker_party": diz})


refined_df.to_csv("data/cleaned_dataset/refined.csv", index=False)
