==> Notebooks/3. Manual Annotation Helper.ipynb <==
import pandas as pd

df=pd.read_csv("data/cleaned_dataset/merged.tsv")

text_id = "ParlaMint-IT_2022-05-25-LEG18-Senato-sed-437.u102"
df[df["ID"] == text_id].text.iloc[0]

#df_m5s = pd.concat([df[df["Speaker_party"] == "M5S.1"], df[df["Speaker_party"] == "M5S.2"]])
#df_scpi = df[df["Speaker_party"] == "SCpI"]
#df_pd = df[df["Speaker_party"] == "PD"]
#df_ncd = df[df["Speaker_party"] == "NCD"]
#df_pdl = df[df["Speaker_party"] == "PdL"]
#df_fi = df[df["Speaker_party"] == "FI-BP"]
#df_lnaut =  pd.concat([df[df["Speaker_party"] == "LN-Aut"], df[df["Speaker_party"] == "L-SP"]])
df_fdi = df[df["Speaker_party"] == "FdI"]

df_ok = df_fdi

df_ok[(df_ok['text'].str.contains("", na=False)) & (df_ok['text_length'] < 8000)].iloc[491]

#id_s = "ParlaMint-IT_2019-10-23-LEG18-Senato-sed-158.u71"
id_s = df_ok[(df_ok['text'].str.contains("", na=False)) & (df_ok['text_length'] < 8000)].iloc[491]["ID"]
df_ok[df_ok["ID"] == id_s].text.iloc[0]
