==> Notebooks/4. Gemini Quality Checker.ipynb <==
import google.generativeai as genai
import pandas as pd
import time
import numpy as np
from scipy import stats
from sklearn.metrics import classification_report
from sklearn.metrics import mean_squared_error, mean_absolute_error

manual_df = pd.read_csv("data/manual_annotated_dataset/manual_data.csv")
cleaned_df = pd.read_csv("data/cleaned_dataset/merged.tsv", sep=",", encoding="utf-8")

merged_df = manual_df.merge(cleaned_df, how='left', left_on='id', right_on="ID")

#Rimuovo i testi usati nel few shot prompt
prompt_text = ["ParlaMint-IT_2014-03-26-LEG17-Senato-sed-217.u153", "ParlaMint-IT_2013-06-25-LEG17-Senato-sed-50.u53", "ParlaMint-IT_2018-06-05-LEG18-Senato-sed-9.u62","ParlaMint-IT_2021-06-09-LEG18-Senato-sed-334.u46"]
ground_df = merged_df.drop(merged_df[merged_df['id'].isin(prompt_text)].index)

#Estraggo i testi da dare in input a gemini
texts_to_classify = []
for id in ground_df["id"]:
    texts_to_classify.append((id, ground_df[ground_df["id"] == id].text.iloc[0]))

### GEN AI

API_KEY = "AIzaSyDJ_uAiYW420dVAyBGjhTHiHu2gpPr8vH4"
genai.configure(api_key=API_KEY)
model = genai.GenerativeModel('gemini-2.0-flash')

def classify_record(text_id, text, prompt, model):
    filled_prompt = prompt.replace('[$$$]',text)
    res = model.generate_content(filled_prompt)
    time.sleep(4)
    return str(text_id + "," +res.text)

#CALCOLO CLASSIFICAZIONI GEMINI
def calcola_classificazioni(texts_to_classify, prompt):
    i = 0
    gemini_classes = []
    for tupla in texts_to_classify:
        res = classify_record(tupla[0], tupla[1], prompt, model)
        gemini_classes.append(res)
        i+=1
        print(i, end = " ")
        
    classification_list = []
    for stringa in gemini_classes:
        stringa = stringa.strip("\n")
        lista = stringa.split(",")
        lista = [lista[0]] + [int(x) for x in lista[1:]]
        classification_list.append(lista)

    classification_df = pd.DataFrame(classification_list, columns = ['id', 'manichean', 'indivisible', 'peoplecentrism', 'antielitism'])
    return classification_df

#CALCOLO CLASSIFICAZIONI GEMINI - CON PROB
def calcola_classificazioni_prob(texts_to_classify, prompt):
    i = 0
    gemini_classes = []
    for tupla in texts_to_classify:
        res = classify_record(tupla[0], tupla[1], prompt, model)
        gemini_classes.append(res)
        i+=1
        print(i, end = " ")
        
    classification_list = []
    for stringa in gemini_classes:
        stringa = stringa.strip("\n")
        lista = stringa.split(",")
        lista = [lista[0]] + [int(x) for x in lista[1:]]
        classification_list.append(lista)

    classification_df = pd.DataFrame(classification_list, columns = ['id', 'manichean', 'indivisible', 'peoplecentrism', 'antielitism', 'prob'])
    return classification_df

def calcola_prestazioni(ground_df, classification_df):
    dimensions = ['manichean', 'indivisible', 'peoplecentrism', 'antielitism']
    for dim in dimensions:
        print("ANALISI: ", dim)
        print("SPEARMAN: ", stats.spearmanr(ground_df[dim], classification_df[dim]))
        print("MSE: ", mean_squared_error(ground_df[dim], classification_df[dim]))
        print("MAE: ", mean_absolute_error(ground_df[dim], classification_df[dim]))
        print(classification_report(ground_df[dim], classification_df[dim]))
        print("\n\n")

### Few Shot

with open('genai/prompt.txt', 'r', encoding="utf8") as f:
    prompt = f.read()

classification_df = pd.read_csv("genai/50_texts_classifications.csv", sep=",", encoding="utf-8")
#classification_df = calcola_classificazioni(texts_to_classify, prompt)

#classification_df.to_csv("genai/50_texts_classifications.csv", index=False)

calcola_prestazioni(ground_df, classification_df)

### Zero Shot

with open('genai/zero-shot-prompt.txt', 'r', encoding="utf8") as f:
    zero_shot_prompt = f.read()

zero_shot_classification = pd.read_csv("genai/zero-shot_50_texts_classifications.csv", sep=",", encoding="utf-8")
#zero_shot_classification = calcola_classificazioni(texts_to_classify, zero_shot_prompt)

#zero_shot_classification.to_csv("genai/zero-shot_50_texts_classifications.csv", index=False)

calcola_prestazioni(ground_df, zero_shot_classification)

### Binary transformation

bin_ground_df = ground_df[['id','manichean', 'indivisible', 'peoplecentrism', 'antielitism']].replace(2, 1)
bin_classification_df = classification_df.replace(2, 1)
bin_zero_classification_df = zero_shot_classification.replace(2, 1)

calcola_prestazioni(bin_ground_df, bin_classification_df)

calcola_prestazioni(bin_ground_df, bin_zero_classification_df)

### Zero-shot 2.0

with open('genai/zero-shot_prompt_2.0.txt', 'r', encoding="utf8") as f:
    zero_shot_prompt_2 = f.read()

zero_shot_classification = pd.read_csv("genai/zero-shot_50_texts_classifications_2.0.csv", sep=",", encoding="utf-8")
#zero_shot_classification_2 = calcola_classificazioni_prob(texts_to_classify, zero_shot_prompt_2)

ground_df.shape

ground_df[ground_df["prob"]>7].shape

zero_shot_classification_2[zero_shot_classification_2["prob"]>8].shape

zero_shot_classification_2.to_csv("genai/zero-shot_50_texts_classifications_2.0.csv", index=False)

calcola_prestazioni(ground_df[ground_df["prob"]>7], zero_shot_classification_2[zero_shot_classification_2["prob"]>7])

### Binary directly

bin_ground_df = ground_df[['id','manichean', 'indivisible', 'peoplecentrism', 'antielitism']].replace(2, 1)

with open('genai/binary_few_prompt.txt', 'r', encoding="utf8") as f:
    binary_few_prompt = f.read()

binary_few_prompt_classification = pd.read_csv("genai/binary_few_classifications.csv", sep=",", encoding="utf-8")
#binary_few_prompt_classification = calcola_classificazioni(texts_to_classify, binary_few_prompt)

#binary_few_prompt_classification.to_csv("genai/binary_few_classifications.csv", index=False)

calcola_prestazioni(bin_ground_df, binary_few_prompt_classification)
