==> Notebooks/7. Create text representations.ipynb <==
import pandas as pd
import sqlite3
import stanza
import spacy
import numpy as np
from sklearn.feature_extraction.text import TfidfVectorizer
from sklearn.decomposition import TruncatedSVD

nlp = spacy.load("it_core_news_sm", disable=["ner", "parser"])

nlp = stanza.Pipeline(lang ='it', processors="tokenize,pos,lemma", use_gpu=True)  # Enable GPU if available

annotated_texts_pd = pd.read_csv("data/annotated_dataset/annotated_texts.csv", sep=",", encoding="utf-8")

## TF-IDF

annotated_texts_df = pd.read_csv("data/annotated_dataset/annotated_texts_repr.csv", sep=",", encoding="utf-8")

vectorizer = TfidfVectorizer()

tfidf_matrix = vectorizer.fit_transform(annotated_texts_df['text'])

n_components = 300
svd = TruncatedSVD(n_components=n_components)
reduced_matrix = svd.fit_transform(tfidf_matrix)

tfidf_final = reduced_matrix.tolist()

annotated_texts_df["tfidf"] = tfidf_final

#annotated_texts_df.to_csv("data/annotated_dataset/annotated_texts_repr.csv", index=False)

## Document Embeddings

sqliteConnection = sqlite3.connect('word_embeddings/itwac128.sqlite')
cursor = sqliteConnection.cursor()
vettori_estratti = {}

def estraiVettore(token):
    if token in vettori_estratti.keys():
        return vettori_estratti[token]
    
    sqlite_select_query = """SELECT * from store WHERE key=?"""
    cursor.execute(sqlite_select_query, (token,))
    record = cursor.fetchall()
    if len(record)>0:
        record = list(record[0])
        vettori_estratti[token] = record[1:-1]
    else:
        return -2
    return record[1:-1] #escludo il primo e l'ultimo elemento (il token e un codice identificativo)

def create_doc_representations(texts_list, pos_option):
    doc_vecs = []
    testi_processati = 0
    for text in texts_list:
        word_vec_list = []
        text_nlp = nlp(text)

        for token in text_nlp:
            if pos_option == False:
                word_vec = estraiVettore(token.lemma_)
            elif token.pos_ in ["NOUN", "VERB", "ADJ"]:
                word_vec = estraiVettore(token.lemma_)
            else:
                continue

            if word_vec == -2:
                continue

            word_vec_list.append(word_vec)

        doc_vec = np.mean(word_vec_list, axis=0).tolist()

        doc_vecs.append(doc_vec)
        testi_processati+=1
        if testi_processati%100 == 0:
            print(testi_processati, end = " ")

    return doc_vecs
    #annotated_texts_pd[column_name] = doc_vecs

texts = annotated_texts_pd.text
doc_vecs = create_doc_representations(texts, False)
annotated_texts_pd["doc_embedding"] = doc_vecs

#### POS Doc Embedding

texts = annotated_texts_pd.text
doc_vecs = create_doc_representations(texts, True)
annotated_texts_pd["doc_embedding_pos"] = doc_vecs

## Linguistic Features

conllu_df = pd.read_csv("data/annotated_dataset/linguistic_profile_conllu.csv", sep="\t", encoding="utf-8")

conllu_df

conllu_df = pd.read_csv("data/annotated_dataset/linguistic_profile_conllu.csv", sep="\t", encoding="utf-8")

conllu_df["Filename"] = conllu_df["Filename"].str.replace("separated_texts/", "")
conllu_df["Filename"] = conllu_df["Filename"].str.replace(".conllu", "")

columns_to_aggregate = list(conllu_df.columns)
columns_to_aggregate.remove("Filename")

conllu_df['linguistic_profile'] = conllu_df[columns_to_aggregate].apply(lambda row: row.tolist(), axis=1)
conllu_df = conllu_df.drop(columns=columns_to_aggregate)

conllu_df

annotated_texts_pd = annotated_texts_pd.merge(conllu_df, how='left', left_on='id', right_on="Filename")
annotated_texts_pd = annotated_texts_pd.drop(columns="Filename")

annotated_texts_pd

annotated_texts_pd.to_csv("data/annotated_dataset/annotated_texts_repr.csv", index=False)
