==> Notebooks/8.1 Traditional Classification.ipynb <==
import pandas as pd
import numpy as np
import lightgbm as lgb
from lightgbm import LGBMClassifier
from sklearn.model_selection import cross_val_score, GridSearchCV
from sklearn.metrics import classification_report
import ast
from sklearn.model_selection import train_test_split
from sklearn.svm import LinearSVC
from sklearn.svm import SVC
from sklearn.preprocessing import StandardScaler
from sklearn.ensemble import RandomForestClassifier, RandomForestRegressor
from sklearn.ensemble import BaggingClassifier
from sklearn.ensemble import AdaBoostClassifier
from sklearn.model_selection import RandomizedSearchCV
from scipy.stats import randint, uniform
from sklearn.utils import resample

annotated_texts_df = pd.read_csv("data/annotated_dataset/annotated_texts_repr.csv", sep=",", encoding="utf-8")

annotated_texts_df.head(2)

def train_fit_multiple_models(X_train, X_test, y_train, y_test):
    
    #LightGBM
    clf_lgbm = LGBMClassifier(random_state=8)
    clf_lgbm.fit(X_train, y_train)
    y_pred = clf_lgbm.predict(X_test)
    print("=======LIGHTGBM: \n", classification_report(y_test ,y_pred))

    #Linear SVC
    clf_linear_svc = LinearSVC(C=1.0, random_state=42)
    clf_linear_svc.fit(X_train, y_train)
    y_pred = clf_linear_svc.predict(X_test)
    print("=======LINEAR SVC: \n",classification_report(y_test, y_pred))
    
    #Non-Linear SVC
    clf_svc = SVC(gamma='auto', C=0.1, kernel='rbf', random_state=42)
    clf_svc.fit(X_train, y_train)
    y_pred = clf_svc.predict(X_test)
    print("=======NON-LINEAR SVC: \n",classification_report(y_test, y_pred))
    
    #Random Forest
    clf_rf = RandomForestClassifier(n_estimators=150, 
                             criterion='gini', 
                             max_depth=None, 
                             min_samples_split=2, 
                             min_samples_leaf=1, 
                             min_weight_fraction_leaf=0.0, 
                             max_features='sqrt', 
                             random_state=0, 
                             n_jobs=-1)
    clf_rf.fit(X_train, y_train)
    y_pred = clf_rf.predict(X_test)
    print("========RANDOM FOREST: \n", classification_report(y_test, y_pred))
    '''
    #Bagging
    clf_bag = BaggingClassifier(estimator=None, n_estimators=100, random_state=0)
    clf_bag.fit(X_train, y_train)
    y_pred = clf_bag.predict(X_test)
    print("========BAGGING: \n", classification_report(y_test, y_pred))
    
    #Bagging SVC
    clf_bag_svc = BaggingClassifier(estimator=SVC(C=1000), n_estimators=10, random_state=0)
    clf_bag_svc.fit(X_train, y_train)
    y_pred = clf_bag_svc.predict(X_test)
    print("========BAGGING SVC: \n", classification_report(y_test, y_pred))
    
    #Bagging RF
    clf_bag_rf = BaggingClassifier(estimator=RandomForestClassifier(n_estimators=100), n_estimators=100, random_state=0)
    clf_bag_rf.fit(X_train, y_train)
    y_pred = clf_bag_rf.predict(X_test)
    print("========BAGGING RF: \n", classification_report(y_test, y_pred))
    '''
    #AdaBoost
    clf_ada = AdaBoostClassifier(estimator=None, n_estimators=100, random_state=0)
    clf_ada.fit(X_train, y_train)
    y_pred = clf_ada.predict(X_test)
    print("========ADABOOST \n", classification_report(y_test, y_pred))
    
    #AdaBoost RF
    clf_ada_rf = AdaBoostClassifier(estimator=RandomForestClassifier(n_estimators=100), n_estimators=100, random_state=0)
    clf_ada_rf.fit(X_train, y_train)
    y_pred = clf_ada_rf.predict(X_test)
    print("========ADABOOST RF \n", classification_report(y_test, y_pred))

def train_fit_best_params(X_train, X_test, y_train, y_test):
    #Light GBM
    param_dist = {
        'n_estimators': randint(100, 1200),                # Range for boosting rounds
        'learning_rate': uniform(0.001, 0.3),              # Learning rate
        'num_leaves': randint(20, 150),                    # Maximum number of leaves in one tree
        'max_depth': randint(3, 15),                       # Maximum depth of the tree
        'min_data_in_leaf': randint(10, 100),              # Minimum number of samples in a leaf
        'feature_fraction': uniform(0.5, 0.5),             # Proportion of features to consider at each split
        'bagging_fraction': uniform(0.5, 0.5),             # Proportion of data to consider at each iteration
        'bagging_freq': randint(1, 10),                    # Frequency of bagging
        'lambda_l1': uniform(0, 5),                        # L1 regularization term
        'lambda_l2': uniform(0, 5)                         # L2 regularization term
    }

    clf_lgb = lgb.LGBMClassifier(verbose=-1, random_state=42)

    random_search = RandomizedSearchCV(
        estimator=clf_lgb,
        param_distributions=param_dist,
        n_iter=50,                   # Number of parameter settings sampled
        scoring='accuracy',
        cv=5,
        verbose=2,
        random_state=42,
        n_jobs=-1
    )

    random_search.fit(X_train, y_train)
    selected_clf =  random_search.best_estimator_
    selected_clf.fit(X_train, y_train)
    y_pred = selected_clf.predict(X_test)
    print("=======BEST PARAM LIGHTGBM:\n", classification_report(y_test ,y_pred))
    
    #Random Forest
    param_dist = {
        'n_estimators': randint(100, 500),                  # Number of trees
        'max_depth': randint(5, 15),                         # Maximum depth of each tree
        'min_samples_split': randint(2, 20),                 # Minimum samples needed to split a node
        'min_samples_leaf': randint(1, 20),                  # Minimum samples needed in each leaf
        'max_features': ['sqrt', 'log2', None],              # Number of features to consider at each split
        'bootstrap': [True, False],                          # Whether to use bootstrapping
        'class_weight': [None, 'balanced']                   # Handle class imbalance
    }

    clf = RandomForestClassifier(random_state=42)

    random_search = RandomizedSearchCV(
        estimator=clf,
        param_distributions=param_dist,
        n_iter=20,                   
        scoring='accuracy',          
        cv=5,
        verbose=2,
        random_state=42,
        n_jobs=-1
    )

    random_search.fit(X_train, y_train)
    selected_clf =  random_search.best_estimator_
    selected_clf.fit(X_train, y_train)
    y_pred = selected_clf.predict(X_test)
    print("=======BEST PARAM RANDOM FOREST:\n", classification_report(y_test ,y_pred))


### Populism

df_pop_train, df_pop_test = train_test_split(annotated_texts_df, test_size=0.25, stratify=annotated_texts_df['pop_sum'], random_state=42)
y_train_pop = df_pop_train["pop_sum"].values
y_test_pop = df_pop_test["pop_sum"].values

#### TF-IDF

X_train = df_pop_train["tfidf"].values
X_test = df_pop_test["tfidf"].values

X_train = np.array([ast.literal_eval(item) for item in X_train])
X_test = np.array([ast.literal_eval(item) for item in X_test])

train_fit_multiple_models(X_train, X_test, y_train_pop, y_test_pop)

train_fit_best_params(X_train, X_test, y_train_pop, y_test_pop)

#### Doc Embedding

X_train = df_pop_train["doc_embedding"].values
X_test = df_pop_test["doc_embedding"].values

X_train = np.array([ast.literal_eval(item) for item in X_train])
X_test = np.array([ast.literal_eval(item) for item in X_test])

train_fit_multiple_models(X_train, X_test, y_train_pop, y_test_pop)

train_fit_best_params(X_train, X_test, y_train_pop, y_test_pop)

#### Doc Embedding POS

X_train = df_pop_train["doc_embedding_pos"].values
X_test = df_pop_test["doc_embedding_pos"].values

X_train = np.array([ast.literal_eval(item) for item in X_train])
X_test = np.array([ast.literal_eval(item) for item in X_test])

train_fit_multiple_models(X_train, X_test, y_train_pop, y_test_pop)

train_fit_best_params(X_train, X_test, y_train_pop, y_test_pop)

#### Linguistic Profile

X_train = df_pop_train["linguistic_profile"].values
X_test = df_pop_test["linguistic_profile"].values

X_train = np.array([ast.literal_eval(item) for item in X_train])
X_test = np.array([ast.literal_eval(item) for item in X_test])

train_fit_multiple_models(X_train, X_test, y_train_pop, y_test_pop)

train_fit_best_params(X_train, X_test, y_train_pop, y_test_pop)

## Polarization

df_pol_train, df_pol_test = train_test_split(annotated_texts_df, test_size=0.25, stratify=annotated_texts_df['polarization'], random_state=42)

y_train_pol = df_pol_train["polarization"].values
y_test_pol = df_pol_test["polarization"].values

#### TF-IDF

X_train = df_pol_train["tfidf"].values
X_test = df_pol_test["tfidf"].values

X_train = np.array([ast.literal_eval(item) for item in X_train])
X_test = np.array([ast.literal_eval(item) for item in X_test])

train_fit_multiple_models(X_train, X_test, y_train_pol, y_test_pol)

train_fit_best_params(X_train, X_test, y_train_pol, y_test_pol)

#### Doc Embedding

X_train = df_pol_train["doc_embedding"].values
X_test = df_pol_test["doc_embedding"].values

X_train = np.array([ast.literal_eval(item) for item in X_train])
X_test = np.array([ast.literal_eval(item) for item in X_test])

train_fit_multiple_models(X_train, X_test, y_train_pol, y_test_pol)

train_fit_best_params(X_train, X_test, y_train_pol, y_test_pol)

#### POS Doc Embedding

X_train = df_pol_train["doc_embedding_pos"].values
X_test = df_pol_test["doc_embedding_pos"].values

X_train = np.array([ast.literal_eval(item) for item in X_train])
X_test = np.array([ast.literal_eval(item) for item in X_test])

train_fit_multiple_models(X_train, X_test, y_train_pol, y_test_pol)

train_fit_best_params(X_train, X_test, y_train_pol, y_test_pol)

#### Linguistic Profile

X_train = df_pol_train["linguistic_profile"].values
X_test = df_pol_test["linguistic_profile"].values

X_train = np.array([ast.literal_eval(item) for item in X_train])
X_test = np.array([ast.literal_eval(item) for item in X_test])

train_fit_multiple_models(X_train, X_test, y_train_pol, y_test_pol)

train_fit_best_params(X_train, X_test, y_train_pol, y_test_pol)
