==> Notebooks/8.2 Traditional Classification - Fine-grained populism.ipynb <==
import pandas as pd
import numpy as np
import lightgbm as lgb
from lightgbm import LGBMClassifier
from sklearn.model_selection import cross_val_score, GridSearchCV
from sklearn.model_selection import cross_val_predict
from sklearn.metrics import classification_report
from sklearn.model_selection import cross_val_score
import ast
from sklearn.ensemble import RandomForestClassifier
from sklearn.ensemble import AdaBoostClassifier
from sklearn.ensemble import BaggingClassifier
from sklearn.svm import SVC
from statistics import mode

annotated_texts_df = pd.read_csv("data/annotated_dataset/annotated_texts_repr.csv", sep=",", encoding="utf-8")

annotated_texts_df.head(1)

x_docembedding = annotated_texts_df["doc_embedding"]
x_docembedding = np.array([ast.literal_eval(item) for item in x_docembedding])

y_manichean = annotated_texts_df["manichean"]
y_peoplecentrism = annotated_texts_df["peoplecentrism"]
y_antielitism = annotated_texts_df["antielitism"]
y_emotional = annotated_texts_df["emotional"]

y_pop_sum = annotated_texts_df["pop_sum"]

clf_lgbm = LGBMClassifier(random_state=8, verbose=-1)

clf_rf = RandomForestClassifier(n_estimators=150, 
                             criterion='gini', 
                             max_depth=None, 
                             min_samples_split=2, 
                             min_samples_leaf=1, 
                             min_weight_fraction_leaf=0.0, 
                             max_features='sqrt', 
                             random_state=0, 
                             n_jobs=-1)

clf_ada_rf = AdaBoostClassifier(estimator=RandomForestClassifier(n_estimators=100), n_estimators=100, random_state=0)

clf_svc = SVC(gamma='auto', C=0.1, kernel='rbf', random_state=42)

model_predictions = []
predictions = []
models = [clf_lgbm,clf_rf,clf_ada_rf,clf_svc]
features = [y_manichean, y_peoplecentrism, y_antielitism, y_emotional]

for feature in features:
    for model in models:
        print(model)
        predictions.append(cross_val_predict(model, x_docembedding, feature, cv=5))
    model_predictions.append(predictions)
    predictions = []
        
model_predictions

mode_list = []
for lista in model_predictions:
    mode_l = [mode(values) for values in zip(*lista)]
    mode_list.append(mode_l)

pop_sum_predictions = [sum(values) for values in zip(*mode_list)]

print(classification_report(y_pop_sum, pop_sum_predictions))

df = pd.DataFrame(model_predictions, columns=["lgbm", "rf", "ada_rf", "svc", "bag"])
df.to_csv("data/classifications/five_models_single_pol.csv", index=False)

len(model_predictions[0])

import csv

with open('data/classifications/manichean_preds.csv', 'w', newline='') as f:
    writer = csv.writer(f)
    writer.writerows(model_predictions[0])
    
with open('data/classifications/peoplecentrism_preds.csv', 'w', newline='') as f:
    writer = csv.writer(f)
    writer.writerows(model_predictions[1])
    
with open('data/classifications/antielitism_preds.csv', 'w', newline='') as f:
    writer = csv.writer(f)
    writer.writerows(model_predictions[2])
    
with open('data/classifications/emotional_preds.csv', 'w', newline='') as f:
    writer = csv.writer(f)
    writer.writerows(model_predictions[3])

features = [y_manichean, y_peoplecentrism, y_antielitism, y_emotional]
predictions = []
for feature in features:
    print(cross_val_score(LGBMClassifier(random_state=8, verbose=-1), x_docembedding, feature, cv=5).mean())
    predict = cross_val_predict(LGBMClassifier(random_state=8, verbose=-1), x_docembedding, feature, cv=5)
    print(classification_report(predict, feature))
