==> Notebooks/8.3 Traditional Classification - Coarse-grained populism.ipynb <==
import pandas as pd
from lightgbm import LGBMClassifier
import numpy as np
import ast
from sklearn.ensemble import RandomForestClassifier
from sklearn.model_selection import cross_val_predict
from sklearn.metrics import classification_report
from sklearn.svm import LinearSVC
from sklearn.ensemble import AdaBoostClassifier

annotated_texts_df = pd.read_csv("data/annotated_dataset/annotated_texts_repr.csv", sep=",", encoding="utf-8")

x_tfidf = annotated_texts_df["tfidf"]
x_tfidf = np.array([ast.literal_eval(item) for item in x_tfidf])

x_docembedding = annotated_texts_df["doc_embedding"]
x_docembedding = np.array([ast.literal_eval(item) for item in x_docembedding])

x_docembedding_pos = annotated_texts_df["doc_embedding_pos"]
x_docembedding_pos = np.array([ast.literal_eval(item) for item in x_docembedding_pos])

x_linguistic_profile = annotated_texts_df["linguistic_profile"]
x_linguistic_profile = np.array([ast.literal_eval(item) for item in x_linguistic_profile])

y_manichean = annotated_texts_df["manichean"]
y_peoplecentrism = annotated_texts_df["peoplecentrism"]
y_antielitism = annotated_texts_df["antielitism"]
y_emotional = annotated_texts_df["emotional"]

y_pop_sum = annotated_texts_df["pop_sum"]

clf_lgbm = LGBMClassifier(random_state=8, verbose=-1)

### Multi feature classification

features = [y_manichean, y_peoplecentrism, y_antielitism, y_emotional]
predictions = []

for feature in features:
    predictions.append(cross_val_predict(clf_lgbm, x_docembedding, feature, cv=5))

predictions

pop_sum_predictions = [sum(values) for values in zip(*predictions)]

pop_sum_coarse = [0 if x < 2 else 1 for x in pop_sum_predictions]

y_pop_sum_coarse = [0 if x < 2 else 1 for x in y_pop_sum]

limit = 2
pop_sum_coarse = [0 if x < limit else 1 for x in pop_sum_predictions]
y_pop_sum_coarse = [0 if x < limit else 1 for x in y_pop_sum]
print(classification_report(pop_sum_coarse, y_pop_sum_coarse))

### Single feature classification

limit = 2
y_pop_sum_coarse = [0 if x < limit else 1 for x in y_pop_sum]

predictions = cross_val_predict(clf_lgbm, x_docembedding, y_pop_sum_coarse, cv=5)
print(classification_report(y_pop_sum_coarse, predictions))

## Ricerca della migliore combinazione modello - rappresentazione (tfidf - linearsvc)

### lightgbm

reprs = [x_tfidf, x_docembedding, x_docembedding_pos, x_linguistic_profile]
for rep in reprs:
    predictions = cross_val_predict(clf_lgbm, rep, y_pop_sum_coarse, cv=5)
    print(classification_report(y_pop_sum_coarse, predictions))

### rf

clf_rf = RandomForestClassifier(n_estimators=150, 
                             criterion='gini', 
                             max_depth=None, 
                             min_samples_split=2, 
                             min_samples_leaf=1, 
                             min_weight_fraction_leaf=0.0, 
                             max_features='sqrt', 
                             random_state=0, 
                             n_jobs=-1)

reprs = [x_tfidf, x_docembedding, x_docembedding_pos, x_linguistic_profile]
for rep in reprs:
    predictions = cross_val_predict(clf_rf, rep, y_pop_sum_coarse, cv=5)
    print(classification_report(y_pop_sum_coarse, predictions))

### svc

clf_linear_svc = LinearSVC(C=1.0, random_state=42)

reprs = [x_tfidf, x_docembedding, x_docembedding_pos, x_linguistic_profile]
for rep in reprs:
    predictions = cross_val_predict(clf_linear_svc, rep, y_pop_sum_coarse, cv=5)
    print(classification_report(y_pop_sum_coarse, predictions))

### adaboost

clf_ada_rf = AdaBoostClassifier(estimator=RandomForestClassifier(n_estimators=100), n_estimators=100, random_state=0)

reprs = [x_tfidf, x_docembedding, x_docembedding_pos, x_linguistic_profile]
for rep in reprs:
    predictions = cross_val_predict(clf_ada_rf, rep, y_pop_sum_coarse, cv=5)
    print(classification_report(y_pop_sum_coarse, predictions))
